rename component name index as default export
